Move BadgeCondition unit tests to separate file
