+++ tests/unit/test_model.py
@@ -1677,68 +1677,3 @@ class Test_Badge_Attributes:
         assert badge.qualified_name == f"badges/global/{CM_ID}.{CM_ATTR_ID}"
         assert badge.badge_metadata_attribute == f"{CM_ID}.{CM_ATTR_ID}"
         assert badge.badge_conditions == [BADGE_CONDITION]
-
-
-class Test_BadgeCondition:
-    @pytest.mark.parametrize(
-        "condition_operator, condition_value, condition_colorhex, message",
-        [
-            (
-                None,
-                "1",
-                BadgeConditionColor.RED,
-                "badge_condition_operator is required",
-            ),
-            (
-                BadgeComparisonOperator.EQ,
-                None,
-                BadgeConditionColor.RED,
-                "badge_condition_value is required",
-            ),
-            (
-                BadgeComparisonOperator.EQ,
-                "1",
-                None,
-                "badge_condition_colorhex is required",
-            ),
-        ],
-    )
-    def test_create_when_required_parameter_is_missing_then_raises_value_error(
-        self, condition_operator, condition_value, condition_colorhex, message
-    ):
-        with pytest.raises(ValueError, match=message):
-            BadgeCondition.create(
-                badge_condition_operator=condition_operator,
-                badge_condition_value=condition_value,
-                badge_condition_colorhex=condition_colorhex,
-            )
-
-    def test_create_with_badge_condition_color(self):
-        condition_operator = BadgeComparisonOperator.EQ
-        condition_value = "1"
-        condition_colorhex = BadgeConditionColor.RED
-
-        sut = BadgeCondition.create(
-            badge_condition_operator=condition_operator,
-            badge_condition_value=condition_value,
-            badge_condition_colorhex=condition_colorhex,
-        )
-
-        assert sut.badge_condition_operator == condition_operator.value
-        assert sut.badge_condition_value == condition_value
-        assert sut.badge_condition_colorhex == condition_colorhex.value
-
-    def test_create_with_badge_condition_color_as_str(self):
-        condition_operator = BadgeComparisonOperator.EQ
-        condition_value = "1"
-        condition_colorhex = "#BF1B1B"
-
-        sut = BadgeCondition.create(
-            badge_condition_operator=condition_operator,
-            badge_condition_value=condition_value,
-            badge_condition_colorhex=condition_colorhex,
-        )
-
-        assert sut.badge_condition_operator == condition_operator.value
-        assert sut.badge_condition_value == condition_value
-        assert sut.badge_condition_colorhex == condition_colorhex
